safari-600.7-branch/LayoutTests/java/lc3/JSObject/ToShort-001-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaScript Object to java.lang.String

PASS dt.setShort(new Boolean(true)); dt.PUB_SHORT is 1
PASS dt.setShort(new Boolean(true)); dt.getShort() is 1
PASS typeof dt.getShort() is "number"
PASS dt.setShort(new Boolean(false)); dt.PUB_SHORT is 0
PASS dt.setShort(new Boolean(false)); dt.getShort() is 0
PASS typeof dt.getShort() is "number"
PASS dt.setShort(new Number(0)); dt.PUB_SHORT is 0
PASS dt.setShort(new Number(0)); dt.getShort() is 0
PASS typeof dt.getShort() is "number"
PASS dt.setShort(new String('32767')); dt.PUB_SHORT is 32767
PASS dt.setShort(new String('32767')); dt.getShort() is 32767
PASS typeof dt.getShort() is "number"
PASS dt.setShort(new String('-32768')); dt.PUB_SHORT is -32768
PASS dt.setShort(new String('-32768')); dt.getShort() is -32768
PASS typeof dt.getShort() is "number"
PASS dt.setShort(new MyObject('5.5')); dt.PUB_SHORT is 5
PASS dt.setShort(new MyObject('5.5')); dt.getShort() is 5
PASS typeof dt.getShort() is "number"
PASS dt.setShort(new MyOtherObject('-107.5')); dt.PUB_SHORT is -107
PASS dt.setShort(new MyOtherObject('-107.5')); dt.getShort() is -107
PASS typeof dt.getShort() is "number"
PASS dt.setShort(new AnotherObject('6666')); dt.PUB_SHORT is 6666
PASS dt.setShort(new AnotherObject('6666')); dt.getShort() is 6666
PASS typeof dt.getShort() is "number"
PASS successfullyParsed is true

TEST COMPLETE
 
